Split VOSpaceClient out of cli for general use
