+++ canfar/cli/vos.py
@@ -20,9 +20,9 @@ from vos import md5_cache
 from vos.vos import CADC_GMS_PREFIX, SortNodeProperty, convert_vospace_time_to_seconds
 
 from canfar import get_logger, set_log_level
-from canfar.client import HTTPClient
 from canfar.hooks.typer.aliases import AliasGroup
 from canfar.utils.console import console
+from canfar.vospace import VOSpaceClient
 
 log = get_logger(__name__)
 
@@ -34,41 +34,6 @@ vos_cli = typer.Typer(
     rich_help_panel="File Management",
     cls=AliasGroup,
 )
-
-
-class VOSpaceClient(HTTPClient):
-    """VOSpace client that inherits authentication from HTTPClient.
-
-    This client automatically uses the active CANFAR authentication context
-    to create an authenticated vos.Client instance.
-    """
-
-    def __init__(self, **kwargs):
-        """Initialize VOSpaceClient with HTTPClient authentication."""
-        super().__init__(**kwargs)
-        self._vos_client = None
-
-    @property
-    def vos_client(self) -> vos.Client:
-        """Get or create authenticated vos.Client instance.
-
-        Returns:
-            vos.Client: Authenticated VOSpace client using the active context's token.
-        """
-        if self._vos_client is None:
-            # Extract token from active authentication context
-            ctx = self.config.context
-
-            # Get access token based on auth mode
-            if hasattr(ctx, 'token') and ctx.token:
-                token = ctx.token.access
-            else:
-                # Fallback for X509 or other auth modes
-                token = None
-
-            self._vos_client = vos.Client(vospace_token=token)
-
-        return self._vos_client
 
 
 # Global flag for human-readable sizes (used by formatting functions)
